Fix withAggregate
